refactor(#1491): split the dump-shape scan out of the test target (campsite rule)
The test target reached 1499 lines against the ~1500 test target, one line
under the file-size ratchet. Moved along a SUBJECT seam into
`support/golden_dump_shapes.rs` (166 lines): the `Unsupported` enum with its
labels and minimal goldens, and `unsupported_shapes`, i.e. everything that
answers "what does this golden CONTAIN" from the dump's own vocabulary. The
test target keeps what it owns — which tables are compared, and which are
excluded for what the scan finds. Test target now 1350 lines.

Two dangling intra-doc links that pointed from the moved code at items still in
the test target are now prose naming them, so no doc claim refers to something
this file cannot see.

Pure move plus visibility: 84 tests pass, unchanged.
